+++ tingting/src/main/java/com/alsif/tingting/concert/entity/Concert.java
@@ -32,7 +32,7 @@ public class Concert {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long seq;
+	private Integer seq;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "concert_hall_seq", nullable = false)
@@ -67,11 +67,11 @@ public class Concert {
 	@OneToMany(mappedBy = "concert")
 	private final List<Grade> grades = new ArrayList<>();
 
-	private Concert(Long seq) {
+	private Concert(Integer seq) {
 		this.seq = seq;
 	}
 
-	public static Concert constructBySeq(Long seq) {
+	public static Concert constructBySeq(Integer seq) {
 		return new Concert(seq);
 	}
